canvas: refactor canvas rewrite
