+++ src/components/Canvas/libs/extractScripts.ts
