+++ src/components/Canvas/libs/loadScripts.ts
@@ -1,34 +1,8 @@
-import { TSrcScript } from "../../../models/script";
-import { TInnerHTMLScript } from "../../P5Canvas/useScriptLoader";
+import { TInnerHTMLScript, TSrcScript } from "../../../models/script";
 import * as R from "ramda";
 
 // investigate https://addyosmani.com/basket.js/ for localstorage caching
 
-export const loadScriptTags = (
-  doc: Document | undefined,
-  scripts: TSrcScript[],
-  setScriptLoaded: (scriptId: string) => void
-) => {
-  if (R.isNil(doc)) return null;
-  for (const scriptProperties of scripts) {
-    const existingScript = doc.getElementById(scriptProperties.id);
-
-    if (!existingScript) {
-      const script = doc.createElement("script");
-      script.id = scriptProperties.id;
-      script.type = "application/javascript";
-      script.async = true;
-      script.src = scriptProperties.path;
-      doc.body.appendChild(script);
-
-      script.onload = () => {
-        setScriptLoaded(scriptProperties.id);
-      };
-    } else {
-      setScriptLoaded(scriptProperties.id);
-    }
-  }
-};
 export const loadScript =
   (doc: Document) =>
   ({ id, path }: TSrcScript) => {
