+++ src/models/script.ts
@@ -2,3 +2,9 @@ export type TSrcScript = {
   path: string;
   id: string;
 };
+
+export type TInnerHTMLScript = {
+  id: string;
+  content: string;
+  shouldOverwrite?: boolean;
+};
